requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile cookbook/llms/groq/research/requirements.in
#
altair==5.3.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   groq
    #   httpx
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.7.0
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   streamlit
    #   typer
distro==1.9.0
    # via groq
exceptiongroup==1.2.1
    # via anyio
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via
    #   phidata
    #   streamlit
groq==0.5.0
    # via -r cookbook/llms/groq/research/requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   groq
    #   phidata
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==24.0
    # via
    #   altair
    #   streamlit
pandas==2.2.2
    # via
    #   altair
    #   streamlit
phidata==2.4.3
    # via -r cookbook/llms/groq/research/requirements.in
pillow==10.3.0
    # via streamlit
protobuf==4.25.3
    # via streamlit
pyarrow==16.0.0
    # via streamlit
pydantic==2.7.0
    # via
    #   groq
    #   phidata
    #   pydantic-settings
pydantic-core==2.18.1
    # via pydantic
pydantic-settings==2.2.1
    # via phidata
pydeck==0.8.1b0
    # via streamlit
pygments==2.17.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   phidata
    #   pydantic-settings
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via phidata
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.4.16
    # via tiktoken
requests==2.31.0
    # via
    #   streamlit
    #   tavily-python
    #   tiktoken
rich==13.7.1
    # via
    #   phidata
    #   streamlit
    #   typer
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   groq
    #   httpx
streamlit==1.33.0
    # via -r cookbook/llms/groq/research/requirements.in
tavily-python==0.3.3
    # via -r cookbook/llms/groq/research/requirements.in
tenacity==8.2.3
    # via streamlit
tiktoken==0.6.0
    # via tavily-python
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via phidata
toolz==0.12.1
    # via altair
tornado==6.4
    # via streamlit
typer==0.12.3
    # via phidata
typing-extensions==4.11.0
    # via
    #   altair
    #   anyio
    #   groq
    #   phidata
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typer
tzdata==2024.1
    # via pandas
dash==2.9.3
dash-bootstrap-components
requests
phi-tools


urllib3==2.2.1
    # via requests
    
